+++ packages/integration_test/lib/integration_test.dart
@@ -14,13 +14,12 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:vm_service/vm_service.dart' as vm;
 
+import 'common.dart';
 import 'src/callback.dart' as driver_actions;
 import 'src/channel.dart';
-import 'src/common.dart';
 import 'src/extension.dart';
 
 export 'src/callback.dart';
-export 'src/common.dart';
 export 'src/extension.dart';
 
 const String _success = 'success';
