+++ packages/integration_test/lib/src/_callback_web.dart
@@ -4,7 +4,7 @@
 
 import 'dart:async';
 
-import 'common.dart';
+import '../common.dart';
 
 /// The dart:html implementation of [CallbackManager].
 ///
