+++ packages/integration_test/test/response_serialization_test.dart
@@ -3,8 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:integration_test/integration_test.dart';
+import 'package:integration_test/common.dart';
 
 void main() {
   test('Serialize and deserialize Failure', () {
